replace index.js
