--- ticket.txt
@@ -1,25 +1,0 @@
-QUESTIONS:
-    i want to start working on documentation and testing type stuff, but not sure that any of this is set in stone. how to balance keeping on top of that annoying work while still developing codebase?
-
-TO DO:
-    Should abandon callbacks in favor of returning Promises (which, confusingly, it also uses in some cases)
-    Should retry data fetches some number of times (say, 3) with exponential backoff
-      *do retry for everything? maybe not posts 
-      *defintiely for get and auth when there are empty responses 
-    Should detect "end of file" (0 byte response) from Sierra and retry (as noted above)
-    Logs should capture any number of retries and when we get an empty response
-
-    Out of scope: setting up new alarms
-    We should support all existing functionality, but may consider changing what things are named to clarify use of the package. This will probably also entail updates to the README, including improving sample
-
-
-DONE? 
-    Should reuse access tokens up to some age, like 5mins (or should reuse an existing token until it's rejected). Currently the package fetches a new one on each call..
-
-    Should retry auth some number of times (say, 3) with exponential backoff
-
-    Should not require integrations to call "promiseAuth" explicitly, but should call it as a new token is needed
-      *wait until theres a 401, and when there is auth error, get new token
-      *retry after auth
-
-
